refactor: update import paths to app.utils.judgment_fetcher
Follow-up to 86864a6 (which captured only the file rename). This commit
contains the actual content changes that update import paths and docstring
references from `app.utils.document_fetcher` to `app.utils.judgment_fetcher`:

- 4 production importers: app/api/schema_generator.py, app/workers.py,
  app/playground.py, juddges_search/retrieval/fetch.py.
- 2 test files: test_judgment_fetcher.py (mocked patch() target strings) and
  test_warning_regressions.py (asserts get_documents_by_id.__module__ +
  function name renamed for consistency).
- 4 deprecation strings in juddges_search/retrieval/fetch.py that pointed
  users to the old module path.
- Module docstring on judgment_fetcher.py freshened (the "legacy naming"
  framing was stale post-rename).

Verified: 11/11 unit tests pass; backend reloads cleanly.

Phase 4 task 4.3a (continuation) of the documents → judgments rename plan.
