+++ backend/tests/app/test_warning_regressions.py
@@ -1,8 +1,8 @@
 """Regression tests for warning-related import paths."""
 
 
-def test_playground_uses_supported_document_fetcher():
-    """Playground should use the maintained document fetcher utility."""
+def test_playground_uses_supported_judgment_fetcher():
+    """Playground should use the maintained judgment fetcher utility."""
     from app import playground
 
-    assert playground.get_documents_by_id.__module__ == "app.utils.document_fetcher"
+    assert playground.get_documents_by_id.__module__ == "app.utils.judgment_fetcher"
